pyproject.toml
[tool.poetry]
name = "IBDCluster"
version = "1.2.2"
description = "A CLI tool to help identify ibd sharing within networks across a locus of interest at biobank scale and then test for phenotypic enrichment within these networks."
authors = ["jtb324 <james.baker@vanderbilt.edu>"]
maintainers = ["jtb324 <james.baker@vanderbilt.edu>"]
readme = "README.md"
homepage = "https://jtb324.github.io/IBDCluster/"
repository = "https://github.com/jtb324/IBDCluster"
keywords = ["python", "clustering", "IBD", "genetics", "relatedness"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
    ]
[tool.poetry.scripts]
drive = "IBDCluster.IBDCluster:app"

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
pandas = "^1.4.0"
numpy = "^1.22.2"
typer = {extras = ["all"], version = "^0.7.0"}
python-dotenv = "^0.19.2"
scipy = "^1.8.0"
tqdm = "^4.62.3"
toml = "^0.10.2"
igraph = "^0.10.4"

[tool.poetry.group.development]
optional=true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^3.0.0"

[tool.poetry.group.development.dependencies]
black = "^22.3.0"
mypy = "^0.931"
ipykernel = "^6.9.1"
data-science-types = "^0.2.23"
pylint = "^2.12.2"
pydantic = "^1.9.0"
pre-commit = "^2.17.0"
perflint = "^0.7.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integtest: marks test as integration test (deselected with '-m \"not integration\"')"
]

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"
